feat: Home page - add routes for home and room - install react router package - Have a separate route for home and room - Idea is to have options to select/create rooms in home
