+++ client/package.json
@@ -14,6 +14,7 @@
     "react-color": "^2.18.1",
     "react-countdown-clock": "^2.7.0",
     "react-dom": "^16.13.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "react-window": "^1.8.5",
     "socket.io-client": "^2.3.0",
