+++ client/src/App.css
@@ -6,111 +6,9 @@ html {
   --buttonColor: linear-gradient(to bottom, #d1e4f6 0%, #c5d8e9 100%);
 }
 
-.App {
-  text-align: center;
-}
-
 body {
   background-color: var(--backgroundColor);
 
   /* Disables pull-to-refresh but allows overscroll glow effects. */
   overscroll-behavior-y: contain;
-}
-
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-.layoutContainer {
-  width: auto;
-  max-height: 100vh;
-}
-
-.userScoreGridItem {
-  margin-left: auto;
-  margin-right: auto;
-  margin-bottom: 10px;
-  margin-top: auto;
-  height: 42vh;
-  display: flex;
-}
-
-.userScoreContainer {
-  width: 100%;
-  display: flex;
-}
-
-.logWindowContainer {
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 20px;
-}
-
-.midPane {
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.midPaneContainer {
-  margin: auto;
-}
-
-.guessBox {
-  width: 100%;
-  background-color: white;
-  z-index: 100;
-}
-
-.canvasContainer {
-  margin: auto;
-}
-
-.inputContainer {
-  position: relative;
-  margin: auto;
-}
-
-.canvasToolboxContainer {
-  margin: auto;
-}
-
-.playersIconContainer {
-  width: inherit;
-  text-align: start;
-  position: absolute;
-  top: 10px;
-  left: 10px;
-}
-
-.playersIconButton {
-  position: relative;
-  z-index: 100;
-}
-
-.playersIcon {
-  border-radius: 50%;
-  background-color: var(--primaryColor);
 }
