+++ client/src/components/home/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import './index.css';
+
+const Home = () => {
+  return <div>Home</div>;
+};
+
+export default Home;
