+++ client/src/components/room/index.css
@@ -1,0 +1,74 @@
+.Room {
+  text-align: center;
+}
+
+.layoutContainer {
+  width: auto;
+  max-height: 100vh;
+}
+
+.userScoreGridItem {
+  margin-left: auto;
+  margin-right: auto;
+  margin-bottom: 10px;
+  margin-top: auto;
+  height: 42vh;
+  display: flex;
+}
+
+.userScoreContainer {
+  width: 100%;
+  display: flex;
+}
+
+.logWindowContainer {
+  margin-left: auto;
+  margin-right: auto;
+  margin-top: 20px;
+}
+
+.midPane {
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.midPaneContainer {
+  margin: auto;
+}
+
+.guessBox {
+  width: 100%;
+  background-color: white;
+  z-index: 100;
+}
+
+.canvasContainer {
+  margin: auto;
+}
+
+.inputContainer {
+  position: relative;
+  margin: auto;
+}
+
+.canvasToolboxContainer {
+  margin: auto;
+}
+
+.playersIconContainer {
+  width: inherit;
+  text-align: start;
+  position: absolute;
+  top: 10px;
+  left: 10px;
+}
+
+.playersIconButton {
+  position: relative;
+  z-index: 100;
+}
+
+.playersIcon {
+  border-radius: 50%;
+  background-color: var(--primaryColor);
+}
